requirements.txt
langchain==0.1.14
gpt4all==2.4.0
chromadb==0.4.24
PyMuPDF==1.24.1
python-dotenv
extract-msg==0.48.5
tabulate==0.9.0
pandoc
pypandoc==1.13
tqdm==4.66.2
sentence_transformers==2.6.1
langchain_community==0.0.31
unstructured
markdown

